+++ lib/android/src/main/java/com/airbnb/android/react/maps/AirMapManager.java
@@ -29,9 +29,13 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
   private static final String REACT_CLASS = "AIRMap";
   private static final int ANIMATE_TO_REGION = 1;
   private static final int ANIMATE_TO_COORDINATE = 2;
-  private static final int FIT_TO_ELEMENTS = 3;
-  private static final int FIT_TO_SUPPLIED_MARKERS = 4;
-  private static final int FIT_TO_COORDINATES = 5;
+  private static final int ANIMATE_TO_VIEWING_ANGLE = 3;
+  private static final int ANIMATE_TO_BEARING = 4;
+  private static final int FIT_TO_ELEMENTS = 5;
+  private static final int FIT_TO_SUPPLIED_MARKERS = 6;
+  private static final int FIT_TO_COORDINATES = 7;
+  private static final int SET_MAP_BOUNDARIES = 8;
+
 
   private final Map<String, Integer> MAP_TYPES = MapBuilder.of(
       "standard", GoogleMap.MAP_TYPE_NORMAL,
@@ -91,6 +95,35 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
     view.map.setMapStyle(new MapStyleOptions(customMapStyleString));
   }
 
+  @ReactProp(name = "mapPadding")
+  public void setMapPadding(AirMapView view, @Nullable ReadableMap padding) {
+    int left = 0;
+    int top = 0;
+    int right = 0;
+    int bottom = 0;
+    double density = (double) view.getResources().getDisplayMetrics().density;
+
+    if (padding != null) {
+      if (padding.hasKey("left")) {
+        left = (int) (padding.getDouble("left") * density);
+      }
+
+      if (padding.hasKey("top")) {
+        top = (int) (padding.getDouble("top") * density);
+      }
+
+      if (padding.hasKey("right")) {
+        right = (int) (padding.getDouble("right") * density);
+      }
+
+      if (padding.hasKey("bottom")) {
+        bottom = (int) (padding.getDouble("bottom") * density);
+      }
+    }
+
+    view.map.setPadding(left, top, right, bottom);
+  }
+
   @ReactProp(name = "showsUserLocation", defaultBoolean = false)
   public void setShowsUserLocation(AirMapView view, boolean showUserLocation) {
     view.setShowsUserLocation(showUserLocation);
@@ -148,6 +181,11 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
     view.map.getUiSettings().setZoomGesturesEnabled(zoomEnabled);
   }
 
+  @ReactProp(name = "zoomControlEnabled", defaultBoolean = true)
+  public void setZoomControlEnabled(AirMapView view, boolean zoomControlEnabled) {
+    view.map.getUiSettings().setZoomControlsEnabled(zoomControlEnabled);
+  }
+
   @ReactProp(name = "rotateEnabled", defaultBoolean = false)
   public void setRotateEnabled(AirMapView view, boolean rotateEnabled) {
     view.map.getUiSettings().setRotateGesturesEnabled(rotateEnabled);
@@ -200,6 +238,8 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
     Double lng;
     Double lngDelta;
     Double latDelta;
+    float bearing;
+    float angle;
     ReadableMap region;
 
     switch (commandId) {
@@ -225,6 +265,18 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
         view.animateToCoordinate(new LatLng(lat, lng), duration);
         break;
 
+      case ANIMATE_TO_VIEWING_ANGLE:
+        angle = (float)args.getDouble(0);
+        duration = args.getInt(1);
+        view.animateToViewingAngle(angle, duration);
+        break;
+
+      case ANIMATE_TO_BEARING:
+        bearing = (float)args.getDouble(0);
+        duration = args.getInt(1);
+        view.animateToBearing(bearing, duration);
+        break;
+
       case FIT_TO_ELEMENTS:
         view.fitToElements(args.getBoolean(0));
         break;
@@ -232,8 +284,13 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
       case FIT_TO_SUPPLIED_MARKERS:
         view.fitToSuppliedMarkers(args.getArray(0), args.getBoolean(1));
         break;
+
       case FIT_TO_COORDINATES:
         view.fitToCoordinates(args.getArray(0), args.getMap(1), args.getBoolean(2));
+        break;
+
+      case SET_MAP_BOUNDARIES:
+        view.setMapBoundaries(args.getMap(0), args.getMap(1));
         break;
     }
   }
@@ -260,17 +317,25 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
 
     return map;
   }
-
-  @Override
+  
   @Nullable
+  @Override
   public Map<String, Integer> getCommandsMap() {
-    return MapBuilder.of(
+    Map<String, Integer> map = MapBuilder.of(
         "animateToRegion", ANIMATE_TO_REGION,
         "animateToCoordinate", ANIMATE_TO_COORDINATE,
+        "animateToViewingAngle", ANIMATE_TO_VIEWING_ANGLE,
+        "animateToBearing", ANIMATE_TO_BEARING,
         "fitToElements", FIT_TO_ELEMENTS,
         "fitToSuppliedMarkers", FIT_TO_SUPPLIED_MARKERS,
         "fitToCoordinates", FIT_TO_COORDINATES
     );
+
+    map.putAll(MapBuilder.of(
+      "setMapBoundaries", SET_MAP_BOUNDARIES
+    ));
+
+    return map;
   }
 
   @Override
@@ -309,7 +374,6 @@ public class AirMapManager extends ViewGroupManager<AirMapView> {
     context.getJSModule(RCTEventEmitter.class)
         .receiveEvent(view.getId(), name, data);
   }
-
 
   @Override
   public void onDropViewInstance(AirMapView view) {